fix(account): scrub failed secret temp files
Fixes #915
